+++ src/Router.tsx
@@ -2,6 +2,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import MainPage from './pages/Main';
 import ShowIpadPage from './pages/ShowIpad';
 import PurchasePage from './pages/Purchase/PurchasePage';
+import AddAccssoryPage from './pages/AddAccessory';
 
 const Router = () => {
   return (
