+++ src/common/Button/AddToCartButton/index.tsx
@@ -4,8 +4,13 @@ import * as S from './AddToCartButton.style';
 export type AddToCartButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & {
   disabled?: boolean;
   children: React.ReactNode;
+  onClick?: () => void;
 };
 
-export const AddToCartButton = ({ children, disabled }: AddToCartButtonProps) => {
-  return <S.Wrapper disabled={disabled}>{children}</S.Wrapper>;
+export const AddToCartButton = ({ children, disabled, onClick }: AddToCartButtonProps) => {
+  return (
+    <S.Wrapper disabled={disabled} onClick={onClick}>
+      {children}
+    </S.Wrapper>
+  );
 };
