+++ src/common/Button/BuyButton/index.tsx
@@ -4,8 +4,13 @@ import * as S from './BuyButton.style';
 export type BuyButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & {
   disabled?: boolean;
   children: React.ReactNode;
+  onClick?: () => void;
 };
 
-export const BuyButton = ({ children, disabled }: BuyButtonProps) => {
-  return <S.Wrapper disabled={disabled}>{children}</S.Wrapper>;
+export const BuyButton = ({ children, disabled, onClick }: BuyButtonProps) => {
+  return (
+    <S.Wrapper disabled={disabled} onClick={onClick}>
+      {children}
+    </S.Wrapper>
+  );
 };
