+++ src/common/Button/DefaultButton/index.tsx
@@ -4,8 +4,13 @@ import * as S from './DefaultButton.style';
 export type DefaultButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & {
   disabled?: boolean;
   children: ReactNode;
+  onClick?: () => void;
 };
 
-export const DefaultButton = ({ children, disabled }: DefaultButtonProps) => {
-  return <S.Wrapper disabled={disabled}>{children}</S.Wrapper>;
+export const DefaultButton = ({ children, disabled, onClick }: DefaultButtonProps) => {
+  return (
+    <S.Wrapper disabled={disabled} onClick={onClick}>
+      {children}
+    </S.Wrapper>
+  );
 };
