+++ src/common/Button/GigaTypeButton/index.tsx
@@ -6,12 +6,19 @@ export type GigaTypeButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & {
   giga: number;
   gigaType: string;
   price: number;
+  onClick?: () => void;
 };
 
-export const GigaTypeButton = ({ disabled, giga, gigaType, price }: GigaTypeButtonProps) => {
+export const GigaTypeButton = ({
+  disabled,
+  giga,
+  gigaType,
+  price,
+  onClick,
+}: GigaTypeButtonProps) => {
   const formattedPrice = new Intl.NumberFormat('ko-KR').format(price);
   return (
-    <S.Wrapper disabled={disabled}>
+    <S.Wrapper disabled={disabled} onClick={onClick}>
       <S.TextWrapper>
         <S.Giga>
           {giga}
