+++ src/common/Button/PlusIconButton/index.tsx
@@ -5,9 +5,10 @@ import { IpadproIcPlusCircleFill, IpadproIcPlusCircleNormal } from '../../../ass
 export type PlusIconButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & {
   disabled?: boolean;
   children: ReactNode;
+  onClick?: () => void;
 };
 
-export const PlusIconButton = ({ children, disabled }: PlusIconButtonProps) => {
+export const PlusIconButton = ({ children, disabled, onClick }: PlusIconButtonProps) => {
   const [isHovered, setIsHovered] = useState(false);
 
   const handleMouseEnter = () => {
@@ -18,7 +19,12 @@ export const PlusIconButton = ({ children, disabled }: PlusIconButtonProps) => {
     setIsHovered(false);
   };
   return (
-    <S.Wrapper disabled={disabled} onMouseEnter={handleMouseEnter} onMouseLeave={handleMouseLeave}>
+    <S.Wrapper
+      disabled={disabled}
+      onMouseEnter={handleMouseEnter}
+      onMouseLeave={handleMouseLeave}
+      onClick={onClick}
+    >
       {children}
       {isHovered ? <IpadproIcPlusCircleFill /> : <IpadproIcPlusCircleNormal />}
     </S.Wrapper>
