+++ src/components/ColumnCarousel/ColumnCarouselWrapper.tsx
@@ -1,13 +1,11 @@
 import { useRef, useState, useEffect } from 'react';
-import { AddToCartButton } from '../../common/Button/AddToCartButton';
 import { ColorCarouselComponent } from './ColorCarouselComponent';
 import * as S from './ColumnCarousel.style';
 import { ConnectCarouselComponent } from './ConnectCarouselComponent';
 import { GigaCarouselComponent } from './GigaCarouselComponent';
 import { ModelCarouselComponent } from './ModelCarouselComponent';
-import { IpadproIcApplebag, IpadproIcScrabNormal, IpadproIcTruck } from '../../assets/icon';
+import { IpadproIcApplebag, IpadproIcTruck } from '../../assets/icon';
 import axios from 'axios';
-import AddAccssoryPage from '../../pages/AddAccessory';
 
 export const ColumnCarouselWrapper = () => {
   const API_URL = import.meta.env.VITE_APP_BASE_URL;
@@ -16,7 +14,6 @@ export const ColumnCarouselWrapper = () => {
   const [startY, setStartY] = useState<number>(0);
   const [scrollTop, setScrollTop] = useState<number>(0);
   const [deliverdate, setDeliverdate] = useState<string>('');
-  const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
 
   const handleMouseDown = (event: React.MouseEvent<HTMLDivElement>) => {
     if (wrapperRef.current) {
@@ -47,10 +44,6 @@ export const ColumnCarouselWrapper = () => {
       wrapperRef.current.scrollBy({ top: delta, behavior: 'smooth' }); // 스크롤 이동
     }
   };
-  const handleModal = () => {
-    setIsModalOpen(true);
-    console.log('클릭');
-  };
   useEffect(() => {
     const fetchData = async () => {
       try {
@@ -68,53 +61,41 @@ export const ColumnCarouselWrapper = () => {
     fetchData();
   }, []);
   return (
-    <>
-      <S.Wrapper
-        ref={wrapperRef}
-        onMouseDown={handleMouseDown}
-        onMouseUp={handleMouseUp}
-        onMouseLeave={handleMouseUp}
-        onMouseMove={handleMouseMove}
-        onWheel={handleScroll}
-      >
-        <ModelCarouselComponent />
-        <ColorCarouselComponent />
-        <GigaCarouselComponent />
-        <ConnectCarouselComponent />
-        <S.PriceWrapper>
-          <S.ItemTitle>11형 iPad Pro Wi-Fi 21GB-스페이스 그레이</S.ItemTitle>
-          <S.FinalPrice>₩1,249,000</S.FinalPrice>
-          <S.VATText>약 ₩285,364의 VAT 포함</S.VATText>
-          <S.CardText>최대 12개월 신용 카드 할부</S.CardText>
-        </S.PriceWrapper>
-        <S.ItemWrapper>
-          <S.ItemBox>
-            <IpadproIcApplebag />
-            <S.ItemText>
-              <S.ItemTitle>픽업 :</S.ItemTitle> <S.ItemSub>재고확인</S.ItemSub>
-            </S.ItemText>
-          </S.ItemBox>
-          <S.ItemBox>
-            <IpadproIcTruck />
-            <S.ItemText>
-              <S.ItemTitle>도착 : </S.ItemTitle>
-              <S.ItemSub>
-                <p>수 {deliverdate} - 무료 배송</p> 추가 배송 옵션 확인
-              </S.ItemSub>
-            </S.ItemText>
-          </S.ItemBox>
-        </S.ItemWrapper>
-        <AddToCartButton onClick={handleModal}>장바구니에 담기</AddToCartButton>
-        <S.LastText>
-          <p>시간이 좀 더 필요하신가요?</p>선택한 기기를 관심 목록에 모두 저장해두고 언제든
-          살펴보던곳부터 다시 이어 보세요.
-          <S.SaveText>
-            <IpadproIcScrabNormal />
-            나중을 위해 저장
-          </S.SaveText>
-        </S.LastText>
-      </S.Wrapper>
-      {isModalOpen && <AddAccssoryPage />}
-    </>
+    <S.Wrapper
+      ref={wrapperRef}
+      onMouseDown={handleMouseDown}
+      onMouseUp={handleMouseUp}
+      onMouseLeave={handleMouseUp}
+      onMouseMove={handleMouseMove}
+      onWheel={handleScroll}
+    >
+      <ModelCarouselComponent />
+      <ColorCarouselComponent />
+      <GigaCarouselComponent />
+      <ConnectCarouselComponent />
+      <S.PriceWrapper>
+        <S.ItemTitle>11형 iPad Pro Wi-Fi 21GB-스페이스 그레이</S.ItemTitle>
+        <S.FinalPrice>₩1,249,000</S.FinalPrice>
+        <S.VATText>약 ₩285,364의 VAT 포함</S.VATText>
+        <S.CardText>최대 12개월 신용 카드 할부</S.CardText>
+      </S.PriceWrapper>
+      <S.ItemWrapper>
+        <S.ItemBox>
+          <IpadproIcApplebag />
+          <S.ItemText>
+            <S.ItemTitle>픽업 :</S.ItemTitle> <S.ItemSub>재고확인</S.ItemSub>
+          </S.ItemText>
+        </S.ItemBox>
+        <S.ItemBox>
+          <IpadproIcTruck />
+          <S.ItemText>
+            <S.ItemTitle>도착 : </S.ItemTitle>
+            <S.ItemSub>
+              <p>수 {deliverdate} - 무료 배송</p> 추가 배송 옵션 확인
+            </S.ItemSub>
+          </S.ItemText>
+        </S.ItemBox>
+      </S.ItemWrapper>
+    </S.Wrapper>
   );
 };
