+++ src/pages/AddAccessory/index.tsx
@@ -8,6 +8,7 @@ import { IcSwatchSilver } from '../../assets/icon';
 import { IcSwatchSpaceGray } from '../../assets/icon';
 import ToggleMenu from '../../components/AddAccessory/toggle';
 import ImgCarousel from '../../components/AddAccessory/ImgCarousel';
+import { useState } from 'react';
 
 const AddAccssoryPage = () => {
   const carouselItems = {
